tst_qfloat16: don't cause UB in converting from float to ints
The test data is provided as float, so we must obey [conv.fpint]/1 when
converting from float to integer types. There are values in the rows
that are outside of the range of short and shorter integers, so just
check the range.

Fixes: QTBUG-125889
Change-Id: If3345151ddf84c43a4f1fffd17d405ee0cd00d44
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit 3c9bb0daaf0cbff9135f88e55bae8243aa937faa)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>