Logging: Introduce a macro to declare an exported Qt logging category
In a follow-up change we will hide Qt's own logging categories in a
namespace. As a result the declaration for Qt's logging categories has
to be different from the declaration for user logging categories.
Introduce a separate macro to declare a logging category exported from
Qt to user code/plugins/tools etc.

Task-number: QTBUG-67692
Change-Id: Ia5ef5225cc0d01e6375e0d2fa1fd7367a5afedc2
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>