Prevent using -std=c++20 on vxworks in tst_qfloat16format
VxWorks 24.03 uses clang 17 that understands the -std=c++20 switch.
However the C++ standard library that is provided by VxWorks does not
support C++20.

Using `CXX_STANDARD 20` leads to -std=c++20 being passed to clang.
Clang in turn enables the preprocessor symbols that are used to check
if C++20 specific features are available (eg __cpp_concepts).

Some files (like qanystringview.h) use such symbols to check if C++20
headers are available.
Since clang enables the symbols, it tries to compile the headers and
hits a compilation error, since the header itself is not available.

Fix the problem by only enabling C++20 when not on VxWorks. This is a
quick fix to enable building tests on CI so it can in turn block future
deliveries that would break VxWorks builds (as right now only the src
dir is built).

Task-number: QTBUG-115777
Change-Id: Idb742927755a2d28c149a28cc0b1290c8ceed7ee
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>