CMake: Assign FindWrapFoo_VERSION variable in QtFindWrapHelper.cmake
The QtFindWrapHelper.cmake file retrieves the version of the
underlying system package version or bundled package version, and
passes that to find_package_handle_standard_args to show which version
was found.
But it did not set a ${PACKAGE_NAME}_VERSION variable, which means
we didn't record the version in qt_find_package, and thus in the final
SBOM.

Make sure we set the variable.

Amends 7ad8c347feb5e336c3833d6d3ac021a0a14b714a

Pick-to: 6.8
Change-Id: Ia0d5b895de36017949e6607d17093459dca302b3
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>